+++ src/services/user/updateSetvice.ts
@@ -1,12 +1,22 @@
+import { InvalidPasswordError } from "@helpers/user-errors/invalidPasswordError";
 import { User } from "../../entities/User";
 import { UserRepository } from "../../repositories/UserRepository";
+import EncryptPasswordService from "./encryptPassword";
 
 
-export class UpdateService {
-  constructor(private readonly userRepository: UserRepository) {}
+export class UpdatePasswordService {
+  constructor(private readonly userRepository: UserRepository) {
+    this.encryptPasswordService = new EncryptPasswordService();
+  }
 
-  async execute(userId: string, userUpdate: Partial<User>): Promise<User> {
-    const updatedUser = await this.userRepository.update(userId, userUpdate);
-    return updatedUser;
+  private readonly encryptPasswordService: EncryptPasswordService;
+
+  async execute(user: User, userId: string, currentPassword: string, newPassword: string): Promise<User> {
+    const isPasswordValid = await this.encryptPasswordService.comparePassword(currentPassword, user.password);
+    if (!isPasswordValid) throw new InvalidPasswordError();
+
+    user.password = await this.encryptPasswordService.encryptPassword(newPassword);
+    await this.userRepository.updatePassword(userId, user.password);
+    return user;
   }
 }
